requirements.txt

streamlit
snowflake-connector-python
pandas
plotly
boto3
snowflake-connector-python
apache-airflow-providers-snowflake
